+++ packages/components/snackbar/src/Snackbar.stories.tsx
@@ -215,22 +215,40 @@ export const QAWithNewLineAction: StoryFn = _args => (
   <div>
     <Snackbar />
 
-    <Button
-      onClick={() =>
-        addSnackbar({
-          message:
-            'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora voluptatum cupiditate ut natus distinctio illum modi, id mollitia sequi dolorem nostrum autem suscipit eius sapiente vitae ipsum amet doloribus praesentium.',
-          actionLabel: 'Undo',
-          onAction: () => console.log('Undone'),
-          icon: <FavoriteFill />,
-          isClosable: true,
-          actionOnNewline: true,
-          timeout: null,
-        })
-      }
-    >
-      Display snackbar
-    </Button>
+    <div className="grid grid-cols-1 gap-xl md:grid-cols-2">
+      <Button
+        onClick={() =>
+          addSnackbar({
+            message:
+              'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora voluptatum cupiditate ut natus distinctio illum modi, id mollitia sequi dolorem nostrum autem suscipit eius sapiente vitae ipsum amet doloribus praesentium.',
+            actionLabel: 'Undo',
+            onAction: () => console.log('Undone'),
+            icon: <FavoriteFill />,
+            isClosable: true,
+            timeout: null,
+          })
+        }
+      >
+        Display snackbar
+      </Button>
+
+      <Button
+        onClick={() =>
+          addSnackbar({
+            message:
+              'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora voluptatum cupiditate ut natus distinctio illum modi, id mollitia sequi dolorem nostrum autem suscipit eius sapiente vitae ipsum amet doloribus praesentium.',
+            actionLabel: 'Undo',
+            onAction: () => console.log('Undone'),
+            icon: <FavoriteFill />,
+            isClosable: true,
+            actionOnNewline: true,
+            timeout: null,
+          })
+        }
+      >
+        Display snackbar with action on new line
+      </Button>
+    </div>
   </div>
 )
 
@@ -244,7 +262,8 @@ export const Action: StoryObj = {
           <Button
             onClick={() =>
               addSnackbar({
-                message: "You're done!",
+                message:
+                  'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora voluptatum cupiditate ut natus distinctio illum.',
                 actionLabel: 'Undo',
                 onAction: () => console.log('Undone'),
               })
@@ -256,7 +275,8 @@ export const Action: StoryObj = {
           <Button
             onClick={() =>
               addSnackbar({
-                message: "You're done! But maybe you should care about what you just did?",
+                message:
+                  'Lorem ipsum dolor sit, amet consectetur adipisicing elit. Tempora voluptatum cupiditate ut natus distinctio illum.',
                 actionLabel: 'Undo',
                 onAction: () => console.log('Undone'),
                 actionOnNewline: true,